chore(array-utils): rename typings file
