+++ array-utils/src/plugins/EnvInfoPlugin.ts
@@ -1,5 +1,5 @@
 import { PluginBase } from "./PluginBase";
-import { EnvDetails } from "../types/typings";
+import { EnvDetails } from "../types/index";
 
 
 export class EnvInfoPlugin extends PluginBase {
