+++ array-utils/src/plugins/UserInfoPlugin.ts
@@ -1,5 +1,5 @@
 import { PluginBase } from "./PluginBase";
-import { UserDetails } from "../types/typings";
+import { UserDetails } from "../types/index";
 import * as os from "node:os";
 
 export class UserInfoPlugin extends PluginBase {
