+++ array-utils/src/types/index.ts
